bench convertBytesToTable
